- Previous solution replaced by new idea based on test class names.
