+++ .github/workflows/check_test_naming_conventions.yml
@@ -1,0 +1,27 @@
+name: Check Test Naming Conventions
+
+on:
+  pull_request:
+    branches: [ master ]
+    types: [ opened, edited, synchronize, reopened ]
+    paths:
+      - 'src/test/scala/**'
+
+jobs:
+  check-test-naming:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4.1.4
+      - name: Set up Python
+        uses: actions/setup-python@v5.1.0
+        with:
+          python-version: '3.x'
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+      - name: Check test naming conventions
+        run: python .github/scripts/check_test_naming.py \
+          ./../../agent/src/test/scala \
+          ./../../database/src/test/scala \
+          ./../../model/src/test/scala \
+          ./../../server/src/test/scala
