+++ build.sbt
@@ -75,26 +75,7 @@ lazy val server = (projectMatrix in file("server"))
       artifactPath / (Compile / packageBin) := baseDirectory.value / s"target/${name.value}-${version.value}.jar",
       testFrameworks += new TestFramework("zio.test.sbt.ZTestFramework"),
       jacocoReportSettings := jacocoSettings(scalaVersion.value, "atum-server"),
-      serverMergeStrategy,
-      commands += Command.command("testIT") { state =>
-        // Apply javaOptions and fork settings only for this command execution
-        val settings = Seq(
-          "set (Test / fork) := true",
-          "set (Test / javaOptions) += \"-DrunIntegration\"",
-          "test"
-        )
-        settings.foldLeft(state)((currentState, setting) => Command.process(setting, currentState))
-      },
-      commands += Command.command("jacocoServer") { state =>
-        // Apply javaOptions and fork settings only for this command execution
-        val settings = Seq(
-          "jacoco",
-          "set (Test / fork) := true",
-          "set (Test / javaOptions) += \"-DrunIntegration\"",
-          "jacoco"
-        )
-        settings.foldLeft(state)((currentState, setting) => Command.process(setting, currentState))
-      }
+      serverMergeStrategy
     ): _*
   )
   .enablePlugins(AssemblyPlugin)
