+++ database/src/test/scala/za/co/absa/atum/database/flows/CreateFlowIntegrationTest.scala
@@ -16,16 +16,15 @@
 
 package za.co.absa.atum.database.flows
 
-import za.co.absa.atum.tags.IntegrationTestTag
 import za.co.absa.balta.DBTestSuite
 
 import scala.util.Random
 
 
-class CreateFlowTest extends DBTestSuite {
+class CreateFlowIntegrationTest extends DBTestSuite {
   private val fncCreateFlow = "flows._create_flow"
 
-  test("Create flow", IntegrationTestTag) {
+  test("Create flow") {
     val partitioningId: Long = Random.nextLong()
     val user = "Geralt of Rivia"
     val flowID = function(fncCreateFlow)
