+++ database/src/test/scala/za/co/absa/atum/database/runs/GetPartitioningAdditionalDataIntegrationTest.scala
@@ -16,11 +16,10 @@
 
 package za.co.absa.atum.database.runs
 
-import za.co.absa.atum.tags.IntegrationTestTag
 import za.co.absa.balta.DBTestSuite
 import za.co.absa.balta.classes.JsonBString
 
-class GetPartitioningAdditionalDataTest extends DBTestSuite{
+class GetPartitioningAdditionalDataIntegrationTest extends DBTestSuite{
 
   private val fncGetPartitioningAdditionalData = "runs.get_partitioning_additional_data"
 
@@ -53,7 +52,7 @@ class GetPartitioningAdditionalDataTest extends DBTestSuite{
       |""".stripMargin
   )
 
-  test("Get partitioning additional data returns additional data for partitioning with additional data", IntegrationTestTag) {
+  test("Get partitioning additional data returns additional data for partitioning with additional data") {
     table("runs.partitionings").insert(
       add("partitioning", partitioning1)
         .add("created_by", "Joseph")
@@ -116,7 +115,7 @@ class GetPartitioningAdditionalDataTest extends DBTestSuite{
 
   }
 
-  test("Get partitioning additional data should return no records for partitioning without additional data", IntegrationTestTag) {
+  test("Get partitioning additional data should return no records for partitioning without additional data") {
     table("runs.partitionings").insert(
       add("partitioning", partitioning2)
         .add("created_by", "Joseph")
@@ -135,7 +134,7 @@ class GetPartitioningAdditionalDataTest extends DBTestSuite{
     }
   }
 
-  test("Get partitioning additional data should return error status code on non existing partitioning", IntegrationTestTag) {
+  test("Get partitioning additional data should return error status code on non existing partitioning") {
     val partitioning = JsonBString(
       """
         |{
