+++ database/src/test/scala/za/co/absa/atum/database/runs/GetPartitioningMeasuresIntegrationTest.scala
@@ -16,15 +16,14 @@
 
 package za.co.absa.atum.database.runs
 
-import za.co.absa.atum.tags.IntegrationTestTag
 import za.co.absa.balta.DBTestSuite
 import za.co.absa.balta.classes.JsonBString
 import za.co.absa.balta.classes.setter.CustomDBType
 
-class GetPartitioningMeasuresTest extends DBTestSuite {
+class GetPartitioningMeasuresIntegrationTest extends DBTestSuite {
   private val fncGetPartitioningMeasures = "runs.get_partitioning_measures"
 
-  test("Get partitioning measures should return partitioning measures for partitioning with measures", IntegrationTestTag) {
+  test("Get partitioning measures should return partitioning measures for partitioning with measures") {
     val partitioning = JsonBString(
       """
         |{
@@ -86,7 +85,7 @@ class GetPartitioningMeasuresTest extends DBTestSuite {
     }
   }
 
-  test("Get partitioning measures should return error status code on non existing partitioning", IntegrationTestTag) {
+  test("Get partitioning measures should return error status code on non existing partitioning") {
     val partitioning = JsonBString(
       """
         |{
@@ -109,7 +108,7 @@ class GetPartitioningMeasuresTest extends DBTestSuite {
       }
   }
 
-  test("Get partitioning measures should return no data status code on partitioning without measures", IntegrationTestTag) {
+  test("Get partitioning measures should return no data status code on partitioning without measures") {
     val partitioning = JsonBString(
       """
         |{
