+++ database/src/test/scala/za/co/absa/atum/database/runs/IsPartitioningValidIntegrationTest.scala
@@ -17,11 +17,10 @@
 package za.co.absa.atum.database.runs
 
 import org.postgresql.util.PSQLException
-import za.co.absa.atum.tags.IntegrationTestTag
 import za.co.absa.balta.DBTestSuite
 import za.co.absa.balta.classes.JsonBString
 
-class IsPartitioningValidTest extends DBTestSuite {
+class IsPartitioningValidIntegrationTest extends DBTestSuite {
 
   private val fncIsPartitioningValid = "validation.is_partitioning_valid"
 
@@ -70,7 +69,7 @@ class IsPartitioningValidTest extends DBTestSuite {
       |""".stripMargin
   )
 
-  test("Partitioning is valid (strict check)", IntegrationTestTag) {
+  test("Partitioning is valid (strict check)") {
    function(fncIsPartitioningValid)
       .setParam("i_partitioning", partitioningValid)
       .execute { queryResult =>
@@ -83,7 +82,7 @@ class IsPartitioningValidTest extends DBTestSuite {
       }
   }
 
-  test("Partitioning is valid but has null values (strict check)", IntegrationTestTag) {
+  test("Partitioning is valid but has null values (strict check)") {
     val caught = intercept[PSQLException](
       function(fncIsPartitioningValid)
         .setParam("i_partitioning", partitioningValidWithNullValues)
@@ -95,7 +94,7 @@ class IsPartitioningValidTest extends DBTestSuite {
     )
   }
 
-  test("Partitioning is valid but has null values (non-strict check)", IntegrationTestTag) {
+  test("Partitioning is valid but has null values (non-strict check)") {
     function(fncIsPartitioningValid)
       .setParam("i_partitioning", partitioningValidWithNullValues)
       .setParam("i_strict_check", false)
@@ -109,7 +108,7 @@ class IsPartitioningValidTest extends DBTestSuite {
       }
   }
 
-  test("Partitioning is invalid", IntegrationTestTag) {
+  test("Partitioning is invalid") {
     val caught = intercept[PSQLException](
       function(fncIsPartitioningValid)
         .setParam("i_partitioning", partitioningInvalid)
