+++ database/src/test/scala/za/co/absa/atum/database/runs/ValidatePartitioningIntegrationTest.scala
@@ -16,11 +16,10 @@
 
 package za.co.absa.atum.database.runs
 
-import za.co.absa.atum.tags.IntegrationTestTag
 import za.co.absa.balta.DBTestSuite
 import za.co.absa.balta.classes.JsonBString
 
-class ValidatePartitioningTest extends DBTestSuite {
+class ValidatePartitioningIntegrationTest extends DBTestSuite {
 
   private val fncValidatePartitioning = "validation.validate_partitioning"
 
@@ -165,7 +164,7 @@ class ValidatePartitioningTest extends DBTestSuite {
       |""".stripMargin
   )
 
-  test("Partitioning is valid (strict check)", IntegrationTestTag) {
+  test("Partitioning is valid (strict check)") {
     function(fncValidatePartitioning)
       .setParam("i_partitioning", partitioningValid)
       .execute { queryResult =>
@@ -173,7 +172,7 @@ class ValidatePartitioningTest extends DBTestSuite {
       }
   }
 
-  test("Partitioning is valid, even when keys have whitespaces (without them they would be duplicated)", IntegrationTestTag) {
+  test("Partitioning is valid, even when keys have whitespaces (without them they would be duplicated)") {
     function(fncValidatePartitioning)
       .setParam("i_partitioning", partitioningValidWhitespacesInKeys)
       .execute { queryResult =>
@@ -181,7 +180,7 @@ class ValidatePartitioningTest extends DBTestSuite {
       }
   }
 
-  test("Partitioning is valid but has null values (strict check)", IntegrationTestTag) {
+  test("Partitioning is valid but has null values (strict check)") {
       function(fncValidatePartitioning)
         .setParam("i_partitioning", partitioningValidWithNullValues)
         .execute { queryResult =>
@@ -195,7 +194,7 @@ class ValidatePartitioningTest extends DBTestSuite {
         }
   }
 
-  test("Partitioning is valid, but one key doesn't have a value (strict check)", IntegrationTestTag) {
+  test("Partitioning is valid, but one key doesn't have a value (strict check)") {
     function(fncValidatePartitioning)
       .setParam("i_partitioning", partitioningValidMissingValue)
       .execute { queryResult =>
@@ -213,7 +212,7 @@ class ValidatePartitioningTest extends DBTestSuite {
       }
   }
 
-  test("Partitioning is valid, but one key doesn't have a value (non-strict check)", IntegrationTestTag) {
+  test("Partitioning is valid, but one key doesn't have a value (non-strict check)") {
     function(fncValidatePartitioning)
       .setParam("i_partitioning", partitioningValidMissingValue)
       .setParam("i_strict_check", false)
@@ -222,7 +221,7 @@ class ValidatePartitioningTest extends DBTestSuite {
       }
   }
 
-  test("Partitioning is valid, but keys don't correspond to values (non-strict check)", IntegrationTestTag) {
+  test("Partitioning is valid, but keys don't correspond to values (non-strict check)") {
     function(fncValidatePartitioning)
       .setParam("i_partitioning", partitioningValidValuesNotCorrespondToKeys)
       .setParam("i_strict_check", false)
@@ -231,7 +230,7 @@ class ValidatePartitioningTest extends DBTestSuite {
       }
   }
 
-  test("Partitioning is valid, but keys don't correspond to values (strict check)", IntegrationTestTag) {
+  test("Partitioning is valid, but keys don't correspond to values (strict check)") {
     function(fncValidatePartitioning)
       .setParam("i_partitioning", partitioningValidValuesNotCorrespondToKeys)
       .execute { queryResult =>
@@ -249,7 +248,7 @@ class ValidatePartitioningTest extends DBTestSuite {
       }
   }
 
-  test("Partitioning is invalid, incompatible version", IntegrationTestTag) {
+  test("Partitioning is invalid, incompatible version") {
     function(fncValidatePartitioning)
       .setParam("i_partitioning", partitioningInvalidIncompatibleVersion)
       .execute { queryResult =>
@@ -263,7 +262,7 @@ class ValidatePartitioningTest extends DBTestSuite {
       }
   }
 
-  test("Partitioning is valid, but it has redundant attribute", IntegrationTestTag) {
+  test("Partitioning is valid, but it has redundant attribute") {
     function(fncValidatePartitioning)
       .setParam("i_partitioning", partitioningValidRedundantAttribute)
       .execute { queryResult =>
@@ -271,7 +270,7 @@ class ValidatePartitioningTest extends DBTestSuite {
       }
   }
 
-  test("Partitioning is invalid, valid JSON but incorrect structure (attribute missing)", IntegrationTestTag) {
+  test("Partitioning is invalid, valid JSON but incorrect structure (attribute missing)") {
     function(fncValidatePartitioning)
       .setParam("i_partitioning", partitioningInvalidBadJSONStructure1)
       .execute { queryResult =>
@@ -285,7 +284,7 @@ class ValidatePartitioningTest extends DBTestSuite {
       }
   }
 
-  test("Partitioning is invalid, duplicated partition keys", IntegrationTestTag) {
+  test("Partitioning is invalid, duplicated partition keys") {
     function(fncValidatePartitioning)
       .setParam("i_partitioning", partitioningInvalidDuplicatedKeys)
       .execute { queryResult =>
@@ -301,7 +300,7 @@ class ValidatePartitioningTest extends DBTestSuite {
       }
   }
 
-  test("Partitioning is invalid, multiple problems", IntegrationTestTag) {
+  test("Partitioning is invalid, multiple problems") {
     function(fncValidatePartitioning)
       .setParam("i_partitioning", partitioningInvalidMultipleProblems)
       .execute { queryResult =>
