--- database/src/test/scala/za/co/absa/atum/tags/IntegrationTestTag.scala
@@ -1,5 +1,0 @@
-package za.co.absa.atum.tags
-
-import org.scalatest.Tag
-
-object IntegrationTestTag extends Tag("IntegrationTest")
