+++ server/src/test/scala/za/co/absa/atum/server/api/http/CreatePartitioningEndpointIntegrationSpec.scala
@@ -18,60 +18,60 @@ package za.co.absa.atum.server.api.http
 
 import org.mockito.Mockito.{mock, when}
 import sttp.client3._
-import sttp.client3.testing.SttpBackendStub
 import sttp.client3.playJson._
+import sttp.client3.testing.SttpBackendStub
 import sttp.model.StatusCode
 import sttp.tapir.server.stub.TapirStubInterpreter
 import sttp.tapir.ztapir.{RIOMonadError, RichZEndpoint}
-import za.co.absa.atum.model.dto.CheckpointDTO
+import za.co.absa.atum.model.dto.AtumContextDTO
 import za.co.absa.atum.server.api.TestData
-import za.co.absa.atum.server.api.controller.CheckpointController
+import za.co.absa.atum.server.api.controller.PartitioningController
 import za.co.absa.atum.server.model.{GeneralErrorResponse, InternalServerErrorResponse}
-import zio.test._
-import zio._
-import za.co.absa.atum.server.model.PlayJsonImplicits.{readsCheckpointDTO, writesCheckpointDTO}
 import zio.test.Assertion.equalTo
+import zio._
+import zio.test._
+import za.co.absa.atum.server.model.PlayJsonImplicits.{readsAtumContextDTO, writesPartitioningSubmitDTO}
 
-object CreateCheckpointEndpointSpec extends ZIOSpecDefault with Endpoints with TestData {
+object CreatePartitioningEndpointIntegrationSpec extends ZIOSpecDefault with Endpoints with TestData {
 
-  private val checkpointControllerMock = mock(classOf[CheckpointController])
+  private val createPartitioningEndpointMock = mock(classOf[PartitioningController])
 
-  when(checkpointControllerMock.createCheckpoint(checkpointDTO1))
-    .thenReturn(ZIO.succeed(checkpointDTO1))
-  when(checkpointControllerMock.createCheckpoint(checkpointDTO2))
+  when(createPartitioningEndpointMock.createPartitioningIfNotExists(partitioningSubmitDTO1))
+    .thenReturn(ZIO.succeed(createAtumContextDTO(partitioningSubmitDTO1)))
+  when(createPartitioningEndpointMock.createPartitioningIfNotExists(partitioningSubmitDTO2))
     .thenReturn(ZIO.fail(GeneralErrorResponse("error")))
-  when(checkpointControllerMock.createCheckpoint(checkpointDTO3))
+  when(createPartitioningEndpointMock.createPartitioningIfNotExists(partitioningSubmitDTO3))
     .thenReturn(ZIO.fail(InternalServerErrorResponse("error")))
 
-  private val checkpointControllerMockLayer = ZLayer.succeed(checkpointControllerMock)
+  private val createPartitioningEndpointMockLayer = ZLayer.succeed(createPartitioningEndpointMock)
 
-  private val createCheckpointServerEndpoint =
-    createCheckpointEndpoint.zServerLogic(CheckpointController.createCheckpoint)
+  private val createPartitioningServerEndpoint =
+    createPartitioningEndpoint.zServerLogic(PartitioningController.createPartitioningIfNotExists)
 
   def spec: Spec[TestEnvironment with Scope, Any] = {
-    val backendStub = TapirStubInterpreter(SttpBackendStub.apply(new RIOMonadError[CheckpointController]))
-      .whenServerEndpoint(createCheckpointServerEndpoint)
+    val backendStub = TapirStubInterpreter(SttpBackendStub.apply(new RIOMonadError[PartitioningController]))
+      .whenServerEndpoint(createPartitioningServerEndpoint)
       .thenRunLogic()
       .backend()
 
     val request = basicRequest
-      .post(uri"https://test.com/api/v1/createCheckpoint")
-      .response(asJson[CheckpointDTO])
+      .post(uri"https://test.com/api/v1/createPartitioning")
+      .response(asJson[AtumContextDTO])
 
-    suite("CreateCheckpointEndpointSuite")(
-      test("Returns expected CheckpointDTO") {
+    suite("CreatePartitioningEndpointIntegrationSuite")(
+      test("Returns expected AtumContextDTO") {
         val response = request
-          .body(checkpointDTO1)
+          .body(partitioningSubmitDTO1)
           .send(backendStub)
 
         val body = response.map(_.body)
         val statusCode = response.map(_.code)
 
-        assertZIO(body <&> statusCode)(equalTo(Right(checkpointDTO1), StatusCode.Created))
+        assertZIO(body <&> statusCode)(equalTo(Right(createAtumContextDTO(partitioningSubmitDTO1)), StatusCode.Ok))
       },
       test("Returns expected BadRequest") {
         val response = request
-          .body(checkpointDTO2)
+          .body(partitioningSubmitDTO2)
           .send(backendStub)
 
         val statusCode = response.map(_.code)
@@ -80,7 +80,7 @@ object CreateCheckpointEndpointSpec extends ZIOSpecDefault with Endpoints with T
       },
       test("Returns expected InternalServerError") {
         val response = request
-          .body(checkpointDTO3)
+          .body(partitioningSubmitDTO3)
           .send(backendStub)
 
         val statusCode = response.map(_.code)
@@ -89,7 +89,7 @@ object CreateCheckpointEndpointSpec extends ZIOSpecDefault with Endpoints with T
       }
     )
   }.provide(
-    checkpointControllerMockLayer
-  ) @@ TestAspect.ifPropNotSet("runIntegration")
+    createPartitioningEndpointMockLayer
+  )
 
 }
